Rename "m4_utilities" to "m4_util" to be consistent with project wide utility module naming.
